# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for LingoDeck — an English learning app",
        "title": "LingoDeck API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/translate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translator"
                ],
                "summary": "Translate text",
                "parameters": [
                    {
                        "description": "Translation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.translateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Translation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorBody"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorBody"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.errorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.errorBody"
                        }
                    }
                }
            }
        },
        "/api/translate/history": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translator"
                ],
                "summary": "Get translation history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Translation"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Translation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "source_lang": {
                    "type": "string"
                },
                "source_text": {
                    "type": "string"
                },
                "target_lang": {
                    "type": "string"
                },
                "target_text": {
                    "type": "string"
                }
            }
        },
...
        "handler.errorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handler.errorDetail"
                }
            }
        },
        "handler.errorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.translateRequest": {
            "type": "object",
            "properties": {
                "source_lang": {
                    "type": "string"
                },
                "target_lang": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}